requirements.txt
fastapi
uvicorn
python-dotenv
pdfplumber
google-generativeai
gradio
